Remove unique_name from service specs
The unique_name parameter was needed because we only had access to the
parameters from {{environment}} and everything under `spec` in the
specs file.

Service templates now have access to two new variables,
`{{service_name}}` and `{{service_instance_name}}`.

In this example we used a `unique_name` parameter, which happened to
be a duplicate value of the service instance name. This parameter was
used for resources such as log group names, which need to be
unique. We can now use `{{service_instance_name}}` instead, without
the need to specify this `unique_name` attribute.

All usages of `{{unique_name}}` in the cfn template file have been
replaced by a combination of `{{service_name}}` and
`{{service_instance_name}}`, except the one for the container name.

Prefixing most usages with `{{service_name}}` will result in these
resources being named `front-end_frontend-dev`. This will be a
recommended approach over only using `{{service_instance_name}}` and
in this case resulting in `frontend-dev` since the same instance name
might used across multiple services. The target group is an exception
since their name does not allow `_`, so we use a double hyphen to
separate the service name from the service instance name.

The container resource does not need this since its parent task
definition is created within a family already named with the
service/service_instance prefix, so it would be redundant to prefix
the container name with the service name. The service instance name is
enough.

---

Important note!

Some resource names, such as for SQS queues and Target Groups have
uniqueness constraints in the region they're created in, while also
having short length limits, respectively 32 and 80 characters.

Using either `{{service_name}}`, `{{service_instance_name}}` or a
combination of the two could result in exceeding these limits. In this
case, the workaround would be to resort to the approach previously
used here with `{{unique_name}}`. For instance, customers would need
to add a `{{target_group_name}}` input and specify its value in the
service specs file for each service instance.

---

This commits addresses the request from
https://github.com/aws/aws-proton-public-roadmap/issues/29